Update patch set 1
Patch Set 1:

> There's no download / fastboot mode on this device?
There is a download mode (shown as HS-USB QDLoader 9006 in device manager) in which the device stays only for 2~3 secs (and then reboots itself). Then there is also a fastboot mode, but requires an usb cable to be plugged in whilst holding both vol buttons.
Should I write instructions for both?

Patch-set: 1
Label: Verified=0