Update patch set 2
Patch Set 2: Code-Review+1

PS2 adds in the current version of the gem to the Gemfile to lock it.  People should test before updating the version of gems.

Note: this gem does not work with jruby.  If we want to conditionalize this gem and conditionally use a different gem for jruby.

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1