Update patch set 1
Patch Set 1:

I said "github" because you are using that as the rendering engine.  But, then why did I say "repo" because that's the path to the approval mechanism.  I guess I could have said I said "github" because that's the path to the approval engine, but I didn't.

I'm in no way attached to any verbage in here.  If you want to apply the steps in here to setup the editing environment and then use this approach to edit this commit, that's fine with me.

Patch-set: 1
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0