Update patch set 3
Patch Set 3:

So far we have instructions for people running linux / macos that uses jekyll as part of their normal build environment and uses gitpages for others.

Before this change, were you using Windows to run jekyll to edit the Wiki?  If so, what were you using for the execjs runtime?

Patch-set: 3