Update patch set 6
Patch Set 6:

I'm not really sure. Zifnab gave me the impression that most modern devices do have the toggle, so I thought it made more sense for it to be opt-out. Ideally, "if present" would mean the user keeps going if they can't find the toggle. But I'll leave it up to you.

Patch-set: 6
Label: Verified=0