Update patch set 2
Patch Set 2:

I think it might be best to add to all motos.  And some kind of note could be added that if your device is not supported by the official moto unlock site you can seek out an alternative like the Sunshine app. Something like that.

Patch-set: 2
Label: Verified=0