Update patch set 3
Patch Set 3:

I have added in a note so this template can work for all but two HTC devices. 

First one to be not included in this request is the HTC 10 (pme). This device has disk encryption, so this method will not apply. 

The other one is the HTC DROID Incredible 4G LTE (fireball), this device is from 2012, HTCDev does not support it.

Patch-set: 3
Label: Verified=0