Update patch set 1
Patch Set 1:

Marios: Yes, sorry about the missing information. I didn't think about it until recently and quickly pushed this change.

Please be aware that the script that was linked is not complete, you have to add your own public keys (see the comments of that change on how to do that) and you also have to change the keys already available in the script.

Patch-set: 1