Update patch set 2
Patch Set 2:

Personally I don't like the usage of smaller text in the TOCs. It makes reading a bit harder than it has to be.

PS: If this is not the place to place such opinion: feel free to totally ignore this.

Patch-set: 2
Reviewer: DDGo <17085@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0