Update recovery instructions for Huawei Honor 5X (kiwi).
The previous instructions for kiwi indicate that volume down and power
should be held. This is not correct, however: the correct way to boot
to recovery is holding volume up and power. This was verified by a
recent user install of LineageOS on kiwi as well as third party
sources.

Change-Id: I5c75f709fccd430ca244446b4fa31b1d4bc7867d
Ref: https://techchomps.com/how-to-enter-recovery-mode-on-the-huawei-honor-5x/