Update patch set 12
Patch Set 12:

Razor, you're right, it is VRUAMDK.

But as for the rest, it is totally correct. The sbl set verifies SW_ID before signature checking, and loading into memory. MDK's SW_ID is 00, every thing above that has a higher increment. Verify with IDA for the functions, and strings for existence of SW_REV and SW_ID. 

To add, do you own the device? Heimdall should not work to install an unsigned recovery.

Patch-set: 12