Update patch set 2
Patch Set 2:

As far as I know TWRP 3.0 and newer doesnt really have problems with flashing LineageOS and Google Apps. As P3100, P3110 and P5110 already have recent versions of TWRP I left out espressowifi because of that and put on explicitly a note about P5100, which doesnt have a up-to-date version.

Please feel to correct me on this if what I'm saying is not true.

Patch-set: 2