Update patch set 1
Patch Set 1:

Ah, is it this one:
http://wiki.lineageos.org/faq.html

From where on the main page is this linked, I could not find a way to navigate from the main wiki page to the faq, should we add it onto the home page for easy finding? There are also questions in my diff not in the link above, can/should I add them in, do they make sense?

Patch-set: 1