Update patch set 1
Patch Set 1: -Code-Review

Actually, the first one was the wrong one, as I used that for reference to fill in the real table which would successfully render (the second one). The long second line can't be avoided unfortunately, as the newlines would break the table.
I also can search through the other pages and switch them to Markdown Tables, too.

Patch-set: 1
Label: Code-Review=0