Update patch set 1
Patch Set 1:

From the perspective of *users*, what value is there to knowing that there is a maintainer if it does not get builds? This just adds hope that some day, maybe, builds will start, which is an invalid assumption. 

From the purposes of bug triage, what purpose is a maintainer field if there are no builds? None, bugs get marked as invalid.

As much as I disagree with a publicly identified maintainer, let this field indicate both that it's good for builds AND $I_M taking responsibility.

Hell, the change was merged from Paul removing maintainers listing from an actively maintained device (d2) because of a showstopper bug. Someone other than Paul felt that was an appropriate change.

Patch-set: 1