Update patch set 1
Patch Set 1:

> From the perspective of *users*, what value is there to knowing
 > that there is a maintainer if it does not get builds? This just
 > adds hope that some day, maybe, builds will start, which is an
 > invalid assumption.

If some user assumes that, we can do nothing about it. The message is clear: "The DEVICE_CODENAME is either unmaintained or does not meet the criteria needed for official LineageOS support status. A build guide is available for developers that would like to make private builds, or even restart official support"

Patch-set: 1