Update patch set 4
Patch Set 4:

This is in the "Developers" section. I like the idea to gain the knowledge not by having to search the web but to have it on the wiki. I did that merging the first time this week and would have been glad if there was a tutorial here. Why not? You guys were beginners as well, right? Did know how to merge a kernel the first time already, did you?

Patch-set: 4
Reviewer: Michael W <13028@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0