Update patch set 15
Patch Set 15:

Other than loki (I'm not sure how we're going to handle, or when Dees_Troy is going to get to looking at it) this is good to go. We will need to do something there, or potentially provide a custom twrp image for users to download in the interim.

Patch-set: 15
Label: Verified=0