Update patch set 1
Patch Set 1:

I dunno tbh. that's not hard to do. I put the note there to sorta distinguish it from fastboot_generic. I suppose it should be easy enough to assume that if you don't have to unlock the bootloader, it's already unlocked. I'm mobile atm, I'll do that when I get back to a computer

Patch-set: 1