Remove invalid download_boot and recovery_boot
There is no key combination that will take you to fastboot or recovery. It is possible to get into "Factory reset" mode using the recovery_boot procedure described currently but that will not take you to recovery. It is possible to go to "Download mode" by holding volume up and then plugging in the USB cable, but that is not fastboot.