Update patch set 1
Patch Set 1:

I understand this example removed feature was not the best - maybe I should have started with themes or the like - but I stand behind the idea of having a list where features that couldn't (yet) be ported to the latest supported versions are listed and explained, so that people can work on them and put them back in.

As an external contributor, and as an user, you have literally no idea what's there already and what is not, unless you specifically remember you used an option that it's not present any more.

This is my take on something that people have been wondering about every release, since years.

Patch-set: 1