Update patch set 5
Patch Set 5:

Yes but to make a simple change I have to got beyond a basic knowledge of git (reason for patchset3, I didn't want it there) I also have to get jekyll running on a local machine if I don't want to wait on github pushes to test.  I also have to look at the templating language to figure out how to make changes.  If this was built on a proper wiki system changes would still be accountable and it would have taken me 5 minutes not the better part of a day to make a simple change.  As of right now the install instructions for the 6P are broken and wont be updated.

Patch-set: 5
Reviewer: Matt Leonberger <17860@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0