Update patch set 1
Patch Set 1:

Clock and timezone are not that important and unrelated, but one can fix it while going through settings.

That app should not be offered in the first place, I agree. However, it is done and noob users install it. Preventing installation of that app, and recurring offerings, makes LineageOS stick to more FOSS.

On IRC, several more people complained about this app begin offered in TWRP.

Patch-set: 1