device_build: Add yasm as a build prerequisite
Some x86 devices, such as Z00A, require the yasm assembler to build.
Add it to the list of prerequisite packages so that builders won't have
to deal with unexpected compile-time errors.

Conditionally insert the package in the list of prerequesites based on
the device's architecture containing the string x86 (this also captures
x86_64).

Change-Id: I2f5bf75e1b2c5afc24838e0b1ed375d272174669