Update patch set 2
Patch Set 2:

This attempt makes it conditional and shows up in the same was as the maven dependency shows up for LineageOS 13.0 builds.

I could try to conditionally insert the package inline with the rest of the generic package dependencies, but I'm afraid that markup would be pretty ugly.

Patch-set: 2