Update patch set 2
Patch Set 2:

In my local test of Michael W's proposal, the package list was rendered incorrectly if the conditional statement evaluated false, due to the insertion of a blank line. The bullet list was broken, and the monospace formatting was completely dropped.

I had to make sure that there was at least one non-blank character on the same line as the conditional statement. (For my test, I arbitrarily chose 5 packages per line. I could have also kept it all on one line; it all washed out when it was rendered.)

-* `bc bison build-essential curl flex g++-multilib gcc-multilib git gnupg gperf imagemagick lib32ncurses5-dev lib32readline-dev lib32z1-dev libesd0-dev liblz4-tool libncurses5-dev libsdl1.2-dev libssl-dev libwxgtk3.0-dev libxml2 libxml2-utils lzop pngcrush rsync schedtool squashfs-tools xsltproc yasm zip zlib1g-dev`
+* `bc bison build-essential curl flex
+g++-multilib gcc-multilib git gnupg gperf
+imagemagick lib32ncurses5-dev lib32readline-dev lib32z1-dev libesd0-dev
+liblz4-tool libncurses5-dev libsdl1.2-dev libssl-dev libwxgtk3.0-dev
+libxml2 libxml2-utils lzop pngcrush rsync
+schedtool squashfs-tools xsltproc {% if device.architecture contains 'x86' %}yasm {% endif %}
+zip zlib1g-dev`

Patch-set: 2
Label: Verified=0