Add `-p` flag to `extract-files.sh`
The wiki described using `extract-files.sh` against the mounted system
image by providing it as the first argument. In testing it appears that
there is now a required `-p` flag for specifying a path on disk. Without
the `-p` `extract-files.sh` will wait for a device to appear in ADB
devices.

Change-Id: I1adfc19c82b0b44b3a2454f97b22291480d91732