wiki: add instruction to log out and back in or reboot before using docker
in order to use docker commands without prefixing sudo, we're adding our user account to the docker group. but for the necessary group membership changes to take effect, one has to log out and back in or reboot. otherwise running docker without `sudo` results in an error like "Got permission denied while trying to connect to the Docker daemon socket (...)".

Change-Id: I6fccd7442e38197eb896c146d9e4859d46d168f8