Update patch set 1
Patch Set 1:

You can probably just cherry-pick that change in your device tree (download->cherry-pick), and push it to gerrit (for which we have a convenient guide in the wiki). Out of curiosity, which device is this for?

Patch-set: 1
Reviewer: Paul Keith <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0