Update patch set 3
Patch Set 3:

Typos can be annoying to fix.

For the record, the band list is from the box of a H990DS.  Band 17 was explicitly listed there, even though FrequencyCheck does not list that band.  I can't help wondering if like the V10 the radio is capable of using additional bands.

Patch-set: 3
Label: Verified=0