Update patch set 3
Patch Set 3:

The vs995/ls997/us996 all have CDMA network as well and this is missing.  The h910/h918 are gsm only. vs995 - VZW, ls997 - sprint, us996 - US unlocked (all gsm/cdma)

ie: for us996:
network:
- tech: 2G, bands: '850 900 1800 1900 MHz GSM; 800 1900 MHz CDMA' }
- {tech: 3G, bands: '850 900 1700/2100 1900 2100 MHz HSDPA; CDMA2000 1xEV-DO' }
- {tech: 4G, bands: '2(1900) 3(1800) 4(1700/2100) 5(850) 7(2600) 12(700) 13(700) 17(700) 20(800) 25(1900)' }

But there does not seem to be a consistency as some devices (S5) are listed "800 1900MHz 1xRTT" for the CDMA 2g portion. or "800 1900MHz EVDO" for the 3G .  The above holds valid for LG devices.  The ls997/vs995/us996 need to be adjusted.

Patch-set: 3