Update patch set 2
Patch Set 2:

(1 comment)

>Why does nobody answer me?

It's only been a couple of days...

This guide was written before the extraction script could automatically extract files from zips. I thought we mentioned somewhere this, but apparently we don't and I think we should.

   extract-files.sh -d *.zip

I'll add the support for the automatic extraction of boot images, the script currently only takes care of /system.

I still think we need to update the instructions for the manual extraction, which is what you did.

Patch-set: 2