Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (2 comments)

So, I've had a look at how the buttons currently are specified among the different YAML device files. It's far from consistent as can be seen on this screenshot: https://i.imgur.com/dV7gJBO.png

Now I think two points are important

1. The specified buttons should be prominent to be immediately visible
2. and – as already mentioned – they should be consistent.

Currently in multiple device files 1) is tried to be achieved by just formatting the buttons `<strong>` (`**`), but I think this is not exactly brilliant as the formatting is the same as for the boot modes (see https://imgur.com/a/pKNZj for an example). Therefore I've decided to

1. use HTML `<kbd>` tags to emphasize the buttons
2. and use `+` as a separator between multiple buttons.

IMHO this looks pretty nice out-of-the-box in the current Jekyll theme. The `+` sign seems to be a common convention to list keystroke/button combos and is in my eyes better than any other listing option because it kind of bundles the buttons together and allows for easier readability when multiple button combos are mentioned in one paragraph. To give you an impression of what I mean:

listing with `,` and `and`: https://i.imgur.com/xPiRdOv.png
listing with `+`: https://i.imgur.com/D2Asrmr.png

I hope everyone is happy with those conventions. I'd be willing to adjust _all_ of the other device files to this formatting in another commit if there are no objections. :)

Thanks,
Salim

Patch-set: 2
Reviewer: Salim B <18013@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0