Update patch set 1
Patch Set 1:

The info like you submitted it is sufficient.

Vendor stock ROM includes the baseband firmware. However, you don't need to install the full stock ROM to get newer bootloader, baseband, etc. - Instead, just the partitions that you need can be flashed.

(However, this is nothing special and works like this on most devices. I just want to point out that "the only way to update a baseband is to install a newer stock ROM" is wrong.)

Patch-set: 1