Update patch set 4
Patch Set 4:

The wiki should reflect the current status of the build roster. I am assuming that the recently removed devices are no longer maintained, based on the fact that no one added them back (after fixing the breaking bugs). Will keep this open for a few days more, to see if any maintainer wants to make any move.

As far as it concerns to wt88047, it was assumed that it was dropped by the maintainer.

Patch-set: 4