Update patch set 1
Patch Set 1:

you are right. will remove it from this change tomorrow, removing the prop and adding validation in another commit to keep this one focussed on minor changes. apart from that please lmk if i did it right for the other properties (ofc i have run the validator and tested a little bit if the output is ok when i modify values around to invalid values, but still...)

Patch-set: 1