Update patch set 1
Patch Set 1: -Code-Review

> (3 comments)
 > 
 > Didnt go through all of them (actually only like the first 5), but
 > you seem to have lost some info during conversion

LE comes standard as of 4.1, EDR comes standard as of 3.0, so I removed instances of features that are part of that versions core spec.

Since this is json-schema under the hood I've been keeping it doublequotes so it plays nice.

What i did miss is that things like 4.0 + LE means that the LE is optionally implemented. so I'm fixing all of those.

Patch-set: 1
Label: Code-Review=0