Update patch set 1
Patch Set 1:

Thank you for looking at it. I'm new to git and gerrit. My idea was to keep the number of templates as it was - currently there is only one template for each manufacturer. I will push a new commit where I created a new template of instructions for these three devices - and a better commit-message :)

Patch-set: 1