Update patch set 3
Patch Set 3:

> 
 > The only way is to use dd command.
 > dogo / odin should be added to necessary changes.
 > Please amend the commit with all the necessary changes.
 > 
 > Thanks for looking into this.

Thank you. I hope I understood you correctly. I have no profound knowledge about all of this. I just try to add a little info to the wiki to help other new users install LineageOs more easily. And here I am learning how to use git and gerrit ;-)

I described the dd-method in the instructions that I commited. The problem is that coming from a stock-rom a new user cannot use dd to install TWRP right away because (if I am not mistaken) it would require root privileges to do so. So I described an intermediate step to first flash TWRP to the "boot" partition first (because "fastboot flash recovery ..." doesn't work on these devices), so you can then "dd"-flash TWRP to "recovery" / FOTA in the next step without having to root the device first. I figured this would be the easiest way to install. If there is a better way that I missed we should prefer it of course.

Have a nice evening (or day) :)

Patch-set: 3
Label: Verified=0