Update patch set 2
Patch Set 2:

> Don't we have some sony devices without recovery partition?

Yes. It might be helpful for this issue: https://review.lineageos.org/#/c/201455/

Unfortunately at the moment the info in the device-pages is incorrect or at least unprecise. For example in "pollux.yml" it says "has_recovery_partition: true" although pollux is among the Sony devices you mentioned.

I read into this in the last couple of days and I believe there is some misunderstanding about what is considered a recovery partition. Initially there wasn't any recovery in these Sony devices. Then Sony added recovery (https://developer.sonymobile.com/2015/05/18/recovery-mode-now-available-for-a-range-of-unlocked-xperia-devices-video/) but handled it differently than other manufacturers. They didn't use a separate partition but created a so called "recovery-in-boot arrangement. This means that the recovery is booted using the regular kernel / boot image in the device." (https://twrp.me/sony/sonyxperiaz1.html).

I would prefer to keep the setting and set it to "false" for the affected Sony devices.  Or maybe remove it from all devices except for those Sony ones. At least this would make it a little bit easier to write separate installation instructions for these devices.

Patch-set: 2
Label: Code-Review=0