Update patch set 1
Patch Set 1: Code-Review+1

Validates without errors. Made assumption that devices which didn't specify "Removable" in the text do have a non-removable battery.
Also, to keep it simple, I made the "tech" one optional so we don't have
to research the tech (if it can be found) for those which didn't have it
until now.
Should be in a reviewable state :)

Patch-set: 1
Label: Code-Review=+1