Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This doesnt look bad so far but actually gives me some generic compile error when I want to test it on github pages. Is it required to add that plugin to the config somehow?

Hi Michael,

thank you for looking into this!

I did not test it with github pages - i ran it locally, where it worked flawlessly.

From the fact that it worked locally i would guess that it is not necessary to add that plugin somewhere (as the plugin file performs a register-call already).

Is it _necessary_ that it runs on github pages, too?

Patch-set: 1