Update patch set 1
Patch Set 1:

The only problem with this is - it seems to break gh pages which we show as an easy way to preview your changes.
Our own instance doesnt run on gh pages so should work and it also works locally on my machine, but maybe you could look into it to make sure to not break it for the users who run gh pages. I didnt get any useful output from it so I can only assume it's the call to "normalize_device_list" which makes it fail, not sure if this can somehow be ifdeffed

Patch-set: 1
Reviewer: Michael W <13028@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0