Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > The only problem with this is - it seems to break gh pages which we show as an easy way to preview your changes.
> > Our own instance doesnt run on gh pages so should work and it also works locally on my machine, but maybe you could look into it to make sure to not break it for the users who run gh pages. I didnt get any useful output from it so I can only assume it's the call to "normalize_device_list" which makes it fail, not sure if this can somehow be ifdeffed
> 
> Ok, i understand.
> 
> I will do some research in the next days to find out how to ifdef this function call away in case the site is run on gh pages.

Ok, i did some research. There does _not_ seem to exist some kind of "please apply filter X if it is available" higher-order filter. Neither in jekyll in general, nor in the list of supported plugins by gh pages.

In the IRC i got an interesting idea of hiding the filter call in markdown and then having the plugin re-reading the result, which would not happen if there was no plugin.
This idea sounds really crazy and scary and i would like to avoid such stuff.

So how do we resolve this situation, Michael?

Patch-set: 1