Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> Jacek, you should have amended your change instead of creating new ones. :)

Sorry, i am completely new to gerrit. I am used to think in the category of "opening a merge request that introduces a series of commits", where it is usual to just append more commits in reaction to reviewer comments.

I can of course squash my commits and push again. Would that fix the situation?

Patch-set: 1