Update patch set 1
Patch Set 1:

Question: Are you willing to update this when we change the formats for some of the properties of the yml? I am asking because we are still adding validation to some of the fields which will most likely change the format for example for storage, sdcard, dimensions (w, h, d).
So either this can go in but needs constant adaption or needs to wait until done (which we dont have a time frame for, yet

Also asking because Im no ruby-guy nor really want to (and only can do so with great struggle) adapt the javascript

Patch-set: 1