Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Question: Are you willing to update this when we change the formats for some of the properties of the yml? I am asking because we are still adding validation to some of the fields which will most likely change the format for example for storage, sdcard, dimensions (w, h, d).
> So either this can go in but needs constant adaption or needs to wait until done (which we dont have a time frame for, yet
> 
> Also asking because Im no ruby-guy nor really want to (and only can do so with great struggle) adapt the javascript

This script will have no problems with the addition of _new_ fields.

Whenever one of the few existing fields gets a new format, this might have to be fixed in the script, of course.
I would argue that this would be rather easy for other people than me, too.

I myself was missing some standardization (i.e. the fields should be structured in a way that it is not necessary to normalize them manually.).
If any changes to the yml format lead to the possibility to completely remove my normalizer plugin, i would certainly welcome that!

Patch-set: 1