Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Regarding the standardisation: We are still adding validation for the various fields which includes standardisation, see https://github.com/LineageOS/lineage_wiki/commits/master/test
> This ofc doesnt get rid of your normalisation as e.g. battery is already a validated/standardized field you seem to have to handle in the script. I cant promise but we are ofc. still working on stuff like sdcards, ram sizes, dimensions

That sounds great! 

So how are we going to handle this now?

Can it be merged and then fixed as soon as the changed arrive, or are we going to wait until the "standardization process" is over?

Patch-set: 1
Reviewer: Jacek Galowicz <18724@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0