Update patch set 3
Patch Set 3:

FWIW this is the (currently) proposed future for addon.d/backuptool for A/B https://review.lineageos.org/#/q/topic:addond-omr1-lineage-15.1

Discussion has begun internally on the merits of this patch series, but it is not yet set in stone or anything. Open to suggestions/improvements on it.

I bring it up because currently the addon.d script that OpenGapps generates is not at all compatible with this patch series. Should we go this direction we either need to not support OpenGapps officially or the addon.d script that OpenGapps creates will have to support this.

Patch-set: 3
Reviewer: Dan Pasanen <3671@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0