Update patch set 1
Patch Set 1:

While we appreciate your effort, in my opinion, this is pointless. You'd have to maintain this whenever a kernel gets updated, not even to mention that different branches can be at different state. For example, N-based CAF kernel for 8996 devices is at 3.18.31, while O-based one is at 3.18.71.

Patch-set: 1