Update patch set 1
Patch Set 1:

my cron can auto push a patch whenever there is a kernel update, but that's not going to be often as the large majority of devices are hundreds of versions behind on versions without upstream support.


lineageos only builds the most recent branch so the script uses the kernel version from the branch with the most recent version.

Patch-set: 1