Update patch set 1
Patch Set 1:

Anything listed on kernel.org is most secure because upstream bug fixes can be merged. anything older than 4.4 is so insecure new device are not permitted to ship with android 8 by google. CVS are not assigned to most bugs so selective patching is a good indicator of an insecure kernel (as demonstrated by Greg Kroah-Hartman)

Patch-set: 1