Update patch set 1
Patch Set 1:

> Abandoned
 > 
 > This feels unnecessary.

Please reconsider; It took me a long time to find this information and one line like this was all I would've needed. If there's a better place to add it let me know, but I was hoping the codename "kltespr" could be mentioned somewhere so that it could help the next person who doesn't trust model numbers.

Patch-set: 1