Update patch set 2
Patch Set 2:

> Abandoned

Hi there. I understand that the idea about a better way to list the button combos was realized in this commit: https://review.lineageos.org/c/196335/ Now it seems super-consistent among all devices. ??

But the initial issue this commit was about remains: There are still no instructions on how to boot directly from download mode to the recovery. These are particulary useful to help users avoid (unintentionally) rebooting into system in between – which results in overwriting the just installed TWRP forces the user to start over again.

Are there any objections to the introduction of the key `recovery_boot_from_download` to the YAML device files and the modification of `recovery_install_heimdall.md` as I've proposed? If not I'd be willing to adapt the commit to the current code base, no problem... :)

Patch-set: 2