Update patch set 1
Patch Set 1:

> > This change is a good idea, but I don't think we should say flash
 > to --BOOT to begin with.  A couple steps later says to make a
 > backup. A backup that doesn't boot. Not good obviously.
 > >
 > 
 > After an Odin flash the device will immediately reboot to system
 > not recovery. Said that I think that flashing TWRP to boot
 > partition is actually a good idea since new users may not know that
 > as soon as the flashing ends they have to press recovery keys. On
 > the backup side, maybe we should tell the users to backup only
 > /data?

No, we really shouldn't be recommending flash recovery to boot. We should describe to uses they need to be prepared to hold the specific key-combo on reboot.

Only having hem, back up /data is not all that useful, because they won't be able to restore said data on a non-stock build.

Patch-set: 1