wiki: Fix LG V20 download boot mode instructions
On many/most LG devices the mode "Download" is readily confused to refer
to a mode where the device rewrites the entire phone image using "LGUP"
to write a "KDZ" file.  Meanwhile the LineageOS Wiki uses the tag
"download_boot" for what is more often referred to as
"fastboot"/"bootloader" mode, which allows writing of arbitrary data to
the recovery or system areas.

Change-Id: I1675b564b8acc57f12e3fd951f647bb2926ca0f9