Update patch set 1
Patch Set 1:

Hello, excuse me if it is out of place for me to reply here. I was crafting a wiki entry for RS988, but then noticed that my commit conflicted with this one. I maintain the only unofficial builds for LOS 14.1 and 15.1 for RS988 on XDA, and I would be happy to help push RS988 to official status for 14.1, which its variants H830 and H850 have.

Patch-set: 1
CC: Evan Muir <19507@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>