Update patch set 6
Patch Set 6:

I think that this is fine, but the only problem is that older devices with no OEM unlock switch are going to this "flag" to be set. Can't we decide whether to show this info based on device's release data? Up to the Samsung guys.

Patch-set: 6
Reviewer: Bruno Martins <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>