Update patch set 2
Patch Set 2:

I just thought that getting USB installed properly on Linux is a task and half and then you get weird pit errors with Heimdall only to discover that the download version or apt-get versions are out of date.  pre-complied versions are 'unofficial'.  So you are forced to compile from source.  It would have saved me some time.  but if you think it's fine to leave as is I'm OK with that.

Patch-set: 2
Reviewer: Gareth Kitchen <18903@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>