Update patch set 2
Patch Set 2: Code-Review+1

ok, after LuK1337 filling me in, my understanding is that if the device is encrypted, either sideloading or twrp is needed. Personally I think that we should instruct users to boot into twrp again, as there is a chance their device is encrypted, and if so, they already have the zips on their device, and sideloading requires additional instructions, and something that the user may be uncomfortable with if they haven't done it before.

Patch-set: 2
Reviewer: Charles IV <18275@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1