Update patch set 4
Patch Set 4: Code-Review-1

> (1 comment)

So, a problem with this is "Wipe" in TWRP when data is selected _won't_ format userdata, it will removed app data & apps, while leaving internal storage intact.

With FBE though, we need to actaully _format_ userdata in initial install cases encrypted at or before 8.0, so we should reccomend "Wipe" ==> "Format Data" in TWRP and explain that Internal Storage will be wiped.

Plus, /cache doesn't exist on a/b, and /system does not (and should not) be wiped. -- as it will wipe the current slot, when in actuality, if anything, we would want to wipe the opposing slot -- which isn't feasible to explain to the average user.

We should really scratch the whole "Wipe stuff while _booted_ into recovery line, and on the second boot into recovery (which is preferably lineage recovery), we will instruct them to "wipe" data -- which will actually format userdata as we need it to be.

Patch-set: 4
Reviewer: Nolen Johnson <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1