Update patch set 2
Patch Set 2: Code-Review-1

Update: decryption is not enabled for the linked twrp recovery image. This is not an issue for ivy since the current installation page for ivy does not link to anywhere, 

but for sumire I am unsure if the previous twrp recovery image had decryption enabled. Any pointers as to how I could figure that out?

I understand if a transition from decryption working to not working is unacceptable, but the current link to the sumire recovery image has no documentation behind it and is severely outdated.

Patch-set: 2
Reviewer: Marco Sirabella <20509@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1