Add special instructions to install Heimdall when required
For some devices, like klte, the minimum version of Heimdall is 1.4.2.
But the packaged version in Ubuntu, in 18.10, is still 1.4.1. Therefore
it's necesary to use a compiled version of Heimdall for the install to work.

It adds a bit of complication in the template, but as Ubuntu one of the main
distributions in the Linux world, I imagine that this kind of advice can
save a good amount of time to the users.

Change-Id: Ie3e0957444139f862ee5c7ebb6a8009518c3c256