Update patch set 3
Patch Set 3: Code-Review-2

I'm fundamentally opposed to this change.

If official TWRP is broken, then a working one should be linked on the wiki. If user is using a TWRP that's not linked on the wiki, that's their problem. If the TWRP linked on the wiki is broken, that's the maintainer's problem and the device should be pulled. Not formatting /data is just asking for problems. "It works fine in most cases" is not the same as "It always works", which is the case for devices that have a proper recovery linked on the wiki and format /data, rather than rm -rf'ing everything but /data/media. Why would we encourage users to do something that isn't going to work all the time? This also adds additional complication and frustration for users who get used to rm -rf rather than format, and then need to change. See how many people in our Slack alone got rekt by A/B not preserving addons with sideload, for example.

Patch-set: 3
Reviewer: Gerrit User 11801 <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2