Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Hi Brinly,
> 
> I understand your frustration totally but I'd like you to revisit your decision later.
> 
> It is actually very good you raised the management issue public. There are many things in Lineage to fix - and leaving the project is definitely not the best way out for everyone.
> 
> I had the same feeling 14.1 has been dropped regardless of opinions of me, syphyr, ShevT and others. We lost ShevT for android_kernel_samsung_jf kernel issues. syphyr has ongoing hard time with gtes2 kernel patches to be accepted too. This and a ton of smaller issues must be addressed.
> 
> I hope Tom, Kevin, Simon will help solving all that shit and stop leaking classy devs like you are.
> 
> Vasyl

This was discussed for months leading up to the official 16.0 release, both amongst the directors and in public channels with other contributors to see and comment. It has always been our custom to maintain official builds for only the latest 2 versions, otherwise storage becomes an issue and build boxes cannot keep up with the build roster. If we kept building legacy branches because a handful of maintainers had interest in it, we would still be building KitKat based ROMs. I'm sorry if you were under the impression that 14.1 would continue forever, but that's just not how it's ever been done nor did anyone ever imply that it would be continued forever. The 14.1 branch is, however, always open to contributions, like 11.0 and 13.0 before it. Please try to keep your issues separate from the discussion at hand; if you have complaints, feel free to take them up with the directors directly - our PMs are always open or you can always reach us at directors@lineageos.org (as stated in the charter).

Patch-set: 3