Update patch set 4
Patch Set 4:

A proper reply will be given later, but for now, I'm on mobile, so this can't easily be fully documented

Firstly, you forgot to remove my BQ project permissions for merge via gerrit - secondly - this issue is definitely not resolved. Amending this commit message is a break of trust in the community again and it really saddens to have seen the directs of this project commit to deepen the distrust to making this abuse of power more public unless an apology is made..

Both - Rashed and Sam have misled and lied to the community by falsely claiming that rights were changed then, when I remember asking for and being granted (by zifnab) permissions in 2018 for my BQ projects. A claim that it was an oversight is clearly a lie (I don't have access to Slack anymore). The community should feel outraged by the Lineage directors willingness to deceive the contributor base which in the end has cost a contributor to this great project.

Patch-set: 4