Update patch set 1
Patch Set 1: Code-Review-2

This is not needed. There are asserts in place and a clear error message is shown to the user in case the minimum required firmware is not installed. That said, the current tip is valid, the latest firmware is always advisable.

Patch-set: 1
Reviewer: Gerrit User 13648 <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2