Update patch set 1
Patch Set 1:

> a clear error message is shown to the user in case the minimum required firmware is not installed. That said, the current tip is valid, the latest firmware is always advisable.

I downloaded 1.8GiB of the latest xiaomi firmware (MI8Global_V10), installed it, then tried to install lineage and saw then that I absolutely need the devel version, which is at 9.5. This was another 1.8GiB download.

I could have saved downloading 2 GiB if it was clear from the beginning that the devel version is absolutely required, even if the version is lower than the latest normal firmware version.

Patch-set: 1